backend/requirements.txt
# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
aiofiles==23.2.1

# ── LangChain core ────────────────────────────────────────
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1

# Optional LLM backends (uncomment the one you need):
# langchain-openai==0.1.8        # OpenAI / Azure OpenAI
# huggingface-hub==0.23.2        # HuggingFace Inference API

# ── Embeddings & Vector Search ────────────────────────────
sentence-transformers==3.0.1     # all-mpnet-base-v2 (downloads on first run, ~420 MB)
faiss-cpu==1.8.0                 # FAISS; falls back to numpy if unavailable
numpy==1.26.4
scikit-learn==1.4.2

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
